--- alembic/versions/bdc4db70b17b_.py
@@ -1,68 +1,0 @@
-"""
-
-Revision ID: bdc4db70b17b
-Revises: b28ff39c02a5
-Create Date: 2021-01-06 23:35:24.867658
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'bdc4db70b17b'
-down_revision = 'b28ff39c02a5'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('bug_status',
-                    sa.Column('id', sa.Integer(), nullable=False),
-                    sa.Column('status', sa.String(length=15), nullable=False),
-                    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-                    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-                    sa.PrimaryKeyConstraint('id')
-                    )
-    op.create_index(op.f('ix_bug_status_id'), 'bug_status', ['id'], unique=True)
-    op.create_table('users',
-                    sa.Column('id', sa.Integer(), nullable=False),
-                    sa.Column('locale', sa.String(length=5), nullable=False),
-                    sa.Column('is_blocked', sa.Boolean(), nullable=True),
-                    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-                    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-                    sa.PrimaryKeyConstraint('id')
-                    )
-    op.create_index(op.f('ix_users_id'), 'users', ['id'], unique=True)
-    op.create_table('bugs',
-                    sa.Column('id', sa.Integer(), nullable=False),
-                    sa.Column('photo_path', sa.String(length=255), nullable=False),
-                    sa.Column('description', sa.Text(), nullable=True),
-                    sa.Column('location', sa.String(length=50), nullable=True),
-                    sa.Column('status', sa.Integer(), nullable=False),
-                    sa.Column('user', sa.Integer(), nullable=True),
-                    sa.Column('cause', sa.Text(length=512), nullable=True),
-                    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-                    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-                    sa.ForeignKeyConstraint(['status'], ['bug_status.id'], ),
-                    sa.ForeignKeyConstraint(['user'], ['users.id'], ),
-                    sa.PrimaryKeyConstraint('id')
-                    )
-    op.create_index(op.f('ix_bugs_id'), 'bugs', ['id'], unique=True)
-    # setup base status
-    op.execute("INSERT INTO bug_status (status) values ('pending');")
-    op.execute("INSERT INTO bug_status (status) values ('registered');")
-    op.execute("INSERT INTO bug_status (status) values ('fixed');")
-    op.execute("INSERT INTO bug_status (status) values ('dropped');")
-    # ### end Alembic commands ###
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_bugs_id'), table_name='bugs')
-    op.drop_table('bugs')
-    op.drop_index(op.f('ix_users_id'), table_name='users')
-    op.drop_table('users')
-    op.drop_index(op.f('ix_bug_status_id'), table_name='bug_status')
-    op.drop_table('bug_status')
-    # ### end Alembic commands ###
