+++ alembic/versions/d1b752c277b6_.py
@@ -1,0 +1,28 @@
+"""
+
+Revision ID: d1b752c277b6
+Revises: b28ff39c02a5
+Create Date: 2021-01-07 15:55:48.348302
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = 'd1b752c277b6'
+down_revision = 'b28ff39c02a5'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('bugs', sa.Column('cause', sa.Text(), nullable=True))
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('bugs', 'cause')
+    # ### end Alembic commands ###
